Fix bug when more than one pre-requisite module is in the same dist.
If a module pre-requisite specifies a version, it is hard to map that
version into a distribution (CPAN package file) version.

If two modules are pre-requisites and are bundled in the same
distribution file one module can override the last module's version
[to require].  But we want the dist's version!  So if there are
multiple modules, the module with the same name as the dist file
has the version that matters.