Use the new makepkg env. var SRCPKGDEST for source packages.
Destination directory logic was revamped. set_pkgdest and get_pkgdest
methods are now less powerful. This is because the logic to guess
where the built package will end up is now more complicated.

SRCPKGDEST is used for source packages, PKGDEST is used for binary
packages. We only know if we are building a source package or a
binary package when create() is called.