Fix removal of redundant dep/makedep.
If a dist depends on a module in build.requires of the META.json, for
example, and also depends on a module in runtime.requires then the
runtime.requires version will always be removed. Internally, CPP lumps
the build.requires list in with the runtime.requires before we get the list
so it is very hard to tell which is which.

A hackish fix is to not remove the runtime requires if the versions
described in the dependencies are not exactly the same.

See Params::Validate for an example of this.